# internal/generated/servers/http/v1/task/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package task

// Ref: #/components/schemas/CreateTaskRequest
type CreateTaskRequest struct {
	Type string `json:"type"`
}

// GetType returns the value of Type.
func (s *CreateTaskRequest) GetType() string {
	return s.Type
}

// SetType sets the value of Type.
func (s *CreateTaskRequest) SetType(val string) {
	s.Type = val
}

// Ref: #/components/schemas/CreateTaskResponse
type CreateTaskResponse struct {
	ID string `json:"id"`
}

// GetID returns the value of ID.
func (s *CreateTaskResponse) GetID() string {
	return s.ID
}

// SetID sets the value of ID.
func (s *CreateTaskResponse) SetID(val string) {
	s.ID = val
}

// Ref: #/components/schemas/DeleteTaskRequest
type DeleteTaskRequest struct {
	ID string `json:"id"`
}

// GetID returns the value of ID.
func (s *DeleteTaskRequest) GetID() string {
	return s.ID
}

// SetID sets the value of ID.
func (s *DeleteTaskRequest) SetID(val string) {
	s.ID = val
}

// Ref: #/components/schemas/DeleteTaskResponse
type DeleteTaskResponse struct {
	Status string `json:"status"`
}

// GetStatus returns the value of Status.
func (s *DeleteTaskResponse) GetStatus() string {
	return s.Status
}

// SetStatus sets the value of Status.
func (s *DeleteTaskResponse) SetStatus(val string) {
	s.Status = val
}

// Ref: #/components/schemas/GetTaskStatusRequest
type GetTaskStatusRequest struct {
	ID string `json:"id"`
}

// GetID returns the value of ID.
func (s *GetTaskStatusRequest) GetID() string {
	return s.ID
}

// SetID sets the value of ID.
func (s *GetTaskStatusRequest) SetID(val string) {
	s.ID = val
}

// Ref: #/components/schemas/GetTaskStatusResponse
type GetTaskStatusResponse struct {
	ID        string `json:"id"`
	Status    string `json:"status"`
	CreatedAt string `json:"createdAt"`
	Duration  string `json:"duration"`
}

// GetID returns the value of ID.
func (s *GetTaskStatusResponse) GetID() string {
	return s.ID
}

// GetStatus returns the value of Status.
func (s *GetTaskStatusResponse) GetStatus() string {
	return s.Status
}

// GetCreatedAt returns the value of CreatedAt.
func (s *GetTaskStatusResponse) GetCreatedAt() string {
	return s.CreatedAt
}

// GetDuration returns the value of Duration.
func (s *GetTaskStatusResponse) GetDuration() string {
	return s.Duration
}

// SetID sets the value of ID.
func (s *GetTaskStatusResponse) SetID(val string) {
	s.ID = val
}

// SetStatus sets the value of Status.
func (s *GetTaskStatusResponse) SetStatus(val string) {
	s.Status = val
}

// SetCreatedAt sets the value of CreatedAt.
func (s *GetTaskStatusResponse) SetCreatedAt(val string) {
	s.CreatedAt = val
}

// SetDuration sets the value of Duration.
func (s *GetTaskStatusResponse) SetDuration(val string) {
	s.Duration = val
}